refactor EndpointParameter and fix linting issues
